+++ tests/fakes.c
@@ -101,7 +101,7 @@ DEFINE_FAKE_VALUE_FUNC(uint16_t, hih8120_getHumidityPercent_x10);
 
 // ---------------------------- ~ MH_Z19 ~ ---------------------------------------
 
-DEFINE_FAKE_VALUE_FUNC(mh_z19_returnCode_t, mh_z19_getCo2Ppm, int *);
+DEFINE_FAKE_VALUE_FUNC(mh_z19_returnCode_t, mh_z19_getCo2Ppm, uint16_t *);
 
 // mh_z19_returnCode_t mh_z19_getCo2Ppm(int *ppm);
 DEFINE_FAKE_VOID_FUNC(mh_z19_initialise, serial_comPort_t);
